+++ profiledef.sh
@@ -16,6 +16,8 @@ bootstrap_tarball_compression=(zstd -c -T0 -19)
 file_permissions=(
   ["/etc/shadow"]="0:0:400"
   ["/root/customize_airootfs.sh"]="0:0:755"
+  ["/usr/bin/kesk"]="0:0:755"
+  ["/usr/lib/kesk/commands/upgrade"]="0:0:755"
   ["/usr/local/bin/"]="0:0:755"
   ["/etc/skel/Desktop/"]="0:0:755"
 )